Changed origin from double to float and fixed IGTL_SONIX_STATUS_HEADER_SIZE
Former-commit-id: 874b11ad1ea1c366f621df219d4260b8a32af626

+++ modules/grabberCommon/cxIGTLinkSonixStatusMessage.h
@@ -78,7 +78,7 @@ protected:
   virtual int  PackBody();
   virtual int  UnpackBody();
 
-  double mDataOrigin[3];
+  float mDataOrigin[3];
   ROI mROI; //Ulterius Region of Interest
   int mNewStatus;
 //  double  mWidth;
